@@ -5,13 +5,13 @@ import 'dart:io';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  /*late*/ ArgResults _results;
+  late ArgResults _results;
   DateTime get from => DateTime.parse(_results['from']);
   DateTime get to => DateTime.parse(_results['to']);
-  bool get showQueries => _results['queries'] /*!*/;
+  bool get showQueries => _results['queries']!;
   int get deltaDays =>
       int.parse(_results['delta'] == null ? '7' : _results['delta']);
-  int get exitCode => _results['help'] ? 0 : null;
+  int? get exitCode => _results['help'] ? 0 : null;
   Options(List<String> args) {
     _parser
       ..addFlag('help',
@@ -35,6 +35,7 @@ class Options {
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
+      exit(-1);
     }
   }
 
@@ -60,25 +61,12 @@ class MeanComputer {
   Duration computeMean(
     List<dynamic> issues,
   ) {
-    bool onlyCustomers = false;
     double sum = 0.0;
-    var count = issues == null ? 0 : issues.length;
+    var count = issues.length;
     if (count == 0) return Duration(seconds: 0);
-    bool hasCustomer = false;
     for (var item in issues) {
       var issue = item as Issue;
       if (issue.closedAt == null) continue;
-      if (onlyCustomers)
-        for (var label in issue.labels.labels) {
-          if (label.label.contains('customer:')) {
-            hasCustomer = true;
-            break;
-          }
-        }
-      if (!onlyCustomers || (onlyCustomers && hasCustomer)) {
-        var delta = issue.closedAt.difference(issue.createdAt);
-        sum += delta.inSeconds;
-      }
     }
     _totalCount += count;
     _totalSumSeconds += sum;
@@ -95,7 +83,7 @@ class MeanComputer {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
