@@ -6,15 +6,15 @@ import 'dart:io';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  /*late*/ ArgResults _results;
-  bool get showClosed => _results['closed'] /*!*/;
-  bool get showMerged => _results['merged'] /*!*/;
+  late ArgResults _results;
+  bool get showClosed => _results['closed']!;
+  bool get showMerged => _results['merged']!;
   bool get onlyNotable => !_results['all-contributors'];
-  bool get authors => _results['authors'] /*!*/;
-  bool get reviewers => _results['reviewers'] /*!*/;
+  bool get authors => _results['authors']!;
+  bool get reviewers => _results['reviewers']!;
   DateTime get from => DateTime.parse(_results.rest[0]);
   DateTime get to => DateTime.parse(_results.rest[1]);
-  int get exitCode => _results['help'] ? 0 : null;
+  int? get exitCode => _results['help'] ? 0 : null;
 
   Options(List<String> args) {
     _parser
@@ -43,16 +43,20 @@ class Options {
       _results = _parser.parse(args);
       if (_results['help']) _printUsage();
       if ((_results['closed'] || _results['merged']) &&
-          _results.rest.length != 2) throw ('need start and end dates!');
+          _results.rest.length != 2)
+        throw ArgParserException('need start and end dates!');
       if (_results['merged'] && _results['closed'])
-        throw ('--merged and --closed are mutually exclusive!');
+        throw ArgParserException(
+            '--merged and --closed are mutually exclusive!');
       if (!_results['authors'] && !_results['reviewers'])
-        throw ('must pass one of --authors or --reviewers!');
+        throw ArgParserException('must pass one of --authors or --reviewers!');
       if (_results['authors'] && _results['reviewers'])
-        throw ('must pass only one of --authors or --reviewers!');
+        throw ArgParserException(
+            'must pass only one of --authors or --reviewers!');
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
+      exit(-1);
     }
   }
 
@@ -67,7 +71,7 @@ class Options {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
 
   // Find the list of folks we're interested in
   final orgMembersContents =
@@ -87,7 +91,7 @@ void main(List<String> args) async {
   final github = GitHub(token);
 
   var state = GitHubIssueState.open;
-  DateRange when = null;
+  DateRange? when = null;
   var rangeType = GitHubDateQueryType.none;
   if (opts.showClosed || opts.showMerged) {
     state = opts.showClosed ? GitHubIssueState.closed : GitHubIssueState.merged;
@@ -130,7 +134,7 @@ void main(List<String> args) async {
   // }
 
   print('There were ${prs.length} pull requests.\n\n');
-  Set<String /*!*/ > allParticipants = Set<String>();
+  Set<String> allParticipants = Set<String>();
   var unpaidContributions = <PullRequest>[];
   var paidContributions = <PullRequest>[];
   for (var item in prs) {
