@@ -6,10 +6,10 @@ import 'dart:io';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  /*late*/ ArgResults _results;
+  late ArgResults _results;
   DateTime get from => DateTime.parse(_results['from']);
   DateTime get to => DateTime.parse(_results['to']);
-  int get exitCode => _results['help'] ? 0 : null;
+  int? get exitCode => _results['help'] ? 0 : null;
 
   Options(List<String> args) {
     _parser
@@ -30,6 +30,7 @@ class Options {
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
+      exit(-1);
     }
   }
 
@@ -39,7 +40,7 @@ class Options {
   }
 }
 
-DateTime nearestSaturday(DateTime /*!*/ when) {
+DateTime? nearestSaturday(DateTime when) {
   var result = when;
   for (int delta = 0; delta < 7; delta++) {
     result = when.add(Duration(days: delta));
@@ -50,7 +51,7 @@ DateTime nearestSaturday(DateTime /*!*/ when) {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
@@ -75,7 +76,7 @@ void main(List<String> args) async {
         issue.createdAt.compareTo(opts.to) <= 0) {
       String key = nearestSaturday(issue.createdAt).toString().substring(0, 10);
       if (!counts.containsKey(key)) counts[key] = 0;
-      counts[key] = counts[key] + 1;
+      counts[key] = counts[key]! + 1;
     }
   });
 
@@ -85,7 +86,7 @@ void main(List<String> args) async {
         issue.createdAt.compareTo(opts.to) <= 0) {
       String key = nearestSaturday(issue.createdAt).toString().substring(0, 10);
       if (!counts.containsKey(key)) counts[key] = 0;
-      counts[key] = counts[key] + 1;
+      counts[key] = counts[key]! + 1;
     }
   });
 
