@@ -6,15 +6,15 @@ import 'dart:io';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  /*late*/ ArgResults _results;
+  late ArgResults _results;
   DateTime get from => DateTime.parse(_results['from']);
   DateTime get to => DateTime.parse(_results['to']);
-  bool get summarize => _results['summarize'] /*!*/;
-  bool get showQueries => _results['queries'] /*!*/;
-  bool get onlyCustomers => _results['customers'] /*!*/;
+  bool get summarize => _results['summarize']!;
+  bool get showQueries => _results['queries']!;
+  bool get onlyCustomers => _results['customers']!;
   int get deltaDays =>
       int.parse(_results['delta'] == null ? '7' : _results['delta']);
-  int get exitCode => _results['help'] ? 0 : null;
+  int? get exitCode => _results['help'] ? 0 : null;
   Options(List<String> args) {
     _parser
       ..addFlag('help',
@@ -46,6 +46,7 @@ class Options {
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
+      exit(-1);
     }
   }
 
@@ -56,7 +57,8 @@ class Options {
   }
 }
 
-int countWithOrWithoutCustomers(List<dynamic> issues, {bool onlyCustomers}) {
+int countWithOrWithoutCustomers(List<dynamic> issues,
+    {required bool onlyCustomers}) {
   var count = issues.length;
   if (onlyCustomers) {
     bool hasCustomer = false;
@@ -88,16 +90,16 @@ class MeanComputer {
 
   MeanComputer();
 
-  Duration meanDurationWithOrWithoutCustomers(List<dynamic> issues,
-      {bool onlyCustomers}) {
+  Duration meanDurationWithOrWithoutCustomers(List<dynamic>? issues,
+      {bool? onlyCustomers}) {
     double sum = 0.0;
     var count = issues == null ? 0 : issues.length;
     if (count == 0) return Duration(seconds: 0);
     bool hasCustomer = false;
-    for (var item in issues) {
+    for (var item in issues!) {
       var issue = item as Issue;
       if (issue.closedAt == null) continue;
-      if (onlyCustomers)
+      if (onlyCustomers!)
         for (var label in issue.labels.labels) {
           if (label.label.contains('customer:')) {
             hasCustomer = true;
@@ -124,7 +126,7 @@ class MeanComputer {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
@@ -212,9 +214,9 @@ void main(List<String> args) async {
       interestingPriorities.forEach((p) => row = '${row}\t${openCount[p]}');
       interestingPriorities.forEach((p) => row = '${row}\t${closeCount[p]}');
       interestingPriorities
-          .forEach((p) => row = '${row}\t${meanUntilClosed[p].inHours}');
+          .forEach((p) => row = '${row}\t${meanUntilClosed[p]!.inHours}');
       interestingPriorities
-          .forEach((p) => row = '${row}\t${meanOpenClosed[p].inHours}');
+          .forEach((p) => row = '${row}\t${meanOpenClosed[p]!.inHours}');
       row = '${row}\t${meanComputerUntilClosed.meanDuration.inHours}';
       row = '${row}\t${meanComputerOpenClosed.meanDuration.inHours}';
       print(row);
