@@ -4,10 +4,10 @@ import 'dart:io';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  /*late*/ ArgResults _results;
+  late ArgResults _results;
   DateTime get from => DateTime.parse(_results['from']);
   DateTime get to => DateTime.parse(_results['to']);
-  int get exitCode => _results['help'] ? 0 : null;
+  int? get exitCode => _results['help'] ? 0 : null;
 
   Options(List<String> args) {
     _parser
@@ -27,6 +27,7 @@ class Options {
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
+      exit(-1);
     }
   }
 
@@ -57,23 +58,25 @@ query {
 }
 
 int extractUniqueUsers(dynamic response) {
-  Set<String /*!*/ > committers = Set<String>();
+  Set<String> committers = Set<String>();
   for (var pr in response['search']['nodes']) {
     committers.add(pr['author']['login']);
   }
   return committers.length;
 }
 
-int extractPullRequestCountResponse(dynamic response) {
+int? extractPullRequestCountResponse(dynamic response) {
   return response['search']['issueCount'];
 }
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
 
   final token = Platform.environment['GITHUB_TOKEN'];
-  final httpLink = HttpLink('https://api.github.com/graphql');
+  final httpLink = HttpLink(
+    uri: 'https://api.github.com/graphql',
+  );
   final auth = AuthLink(
     getToken: () async => 'Bearer $token',
   );
@@ -86,6 +89,7 @@ void main(List<String> args) async {
     final q = makeQuery(start, until);
     final options = QueryOptions(document: gql(q));
     final result = await client.query(options);
+    final timeStampFrom = start.toString().substring(0, 10);
     final timeStampTo = until.toIso8601String().substring(0, 10);
 
     if (result.hasException) {
