@@ -5,10 +5,10 @@ import 'dart:io';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  /*late*/ ArgResults _results;
+  late ArgResults _results;
   DateTime get from => DateTime.parse(_results['from']);
   DateTime get to => DateTime.parse(_results['to']);
-  int get exitCode => _results['help'] ? 0 : null;
+  int? get exitCode => _results['help'] ? 0 : null;
   Options(List<String> args) {
     _parser
       ..addFlag('help',
@@ -28,6 +28,7 @@ class Options {
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
+      exit(-1);
     }
   }
 
@@ -55,7 +56,7 @@ void printHeader(Options opts, String which) {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
@@ -75,9 +76,9 @@ void main(List<String> args) async {
     labels: ['P0'],
   );
 
-  var open = <Issue /*!*/ >[];
-  var openedThisPeriod = <Issue /*!*/ >[];
-  var closedThisPeriod = <Issue /*!*/ >[];
+  var open = <Issue>[];
+  var openedThisPeriod = <Issue>[];
+  var closedThisPeriod = <Issue>[];
 
   openIssues.forEach((issue) {
     if (issue.state == "OPEN") open.add(issue);
