@@ -19,11 +19,11 @@ final _client = GraphQLClient(cache: GraphQLCache(), link: _link);
 
 /// Represents a page of information from GitHub.
 class PageInfo {
-  final String _startCursor;
+  final String? _startCursor;
   get startCursor => _startCursor;
-  final bool _hasNextPage;
+  final bool? _hasNextPage;
   get hasNextPage => _hasNextPage;
-  final String _endCursor;
+  final String? _endCursor;
   get endCursor => _endCursor;
   PageInfo(this._startCursor, this._endCursor, this._hasNextPage);
   static PageInfo fromGraphQL(dynamic node) {
@@ -68,7 +68,7 @@ class Reaction {
     "THUMBS_UP"
   ];
 
-  final String _content;
+  final String? _content;
   get content => _content;
   bool get positive =>
       _content == "HEART" || _content == "HOORAY" || _content == "THUMBS_UP";
@@ -82,7 +82,7 @@ class Reaction {
   }
 
   @override
-  String toString() => _content;
+  String toString() => _content!;
 
   @override
   bool operator ==(Object other) =>
@@ -102,20 +102,20 @@ class Reaction {
 }
 
 class Comment {
-  final Actor _author;
+  final Actor? _author;
   get author => _author;
-  final DateTime _createdAt;
+  final DateTime? _createdAt;
   get createdAt => _createdAt;
-  final String _body;
+  final String? _body;
   get body => _body;
-  final String _id;
+  final String? _id;
   get id => _id;
   get reactionStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = _reactionQuery
-          .replaceAll(r'${id}', this._id)
+          .replaceAll(r'${id}', this._id!)
           .replaceAll(r'${after}', after);
       final options = QueryOptions(document: gql(query));
 
@@ -127,7 +127,7 @@ class Comment {
       }
       try {
         PageInfo pageInfo =
-            PageInfo.fromGraphQL(page.data['node']['reactions']['pageInfo']);
+            PageInfo.fromGraphQL(page.data!['node']['reactions']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
@@ -136,7 +136,7 @@ class Comment {
       // Parse the responses into a buffer
       var bufferReactions = <Reaction>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['node']['reactions']['nodes']) {
+      for (var jsonSub in page.data!['node']['reactions']['nodes']) {
         bufferReactions.add(Reaction.fromGraphQL(jsonSub));
       }
 
@@ -145,7 +145,7 @@ class Comment {
         do {
           yield bufferReactions[bufferIndex++];
         } while (bufferIndex < bufferReactions.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   Comment(this._author, this._id, this._createdAt, this._body);
@@ -198,7 +198,7 @@ class Comment {
 }
 
 class Label {
-  final String _label;
+  final String? _label;
   get label => _label;
   Label(this._label);
   static Label fromGraphQL(dynamic node) {
@@ -206,7 +206,7 @@ class Label {
   }
 
   @override
-  String toString() => _label;
+  String toString() => _label!;
 
   @override
   bool operator ==(Object other) =>
@@ -231,7 +231,7 @@ class Labels {
   get length => _labels.length;
   void append(l) => _labels.add(l);
   bool contains(l) => _labels.contains(l);
-  bool containsString(s) => labels.contains(Label(s));
+  bool? containsString(s) => labels.contains(Label(s));
   bool intersect(List<Label> list) {
     for (var l in list) if (_labels.contains(l)) return true;
     return false;
@@ -263,7 +263,7 @@ class Labels {
     final priorities = {'P0', 'P1', 'P2', 'P3', 'P4', 'P5', 'P6'};
     String result = '';
     priorities.forEach((p) {
-      if (containsString(p)) {
+      if (containsString(p)!) {
         result = p;
         return;
       }
@@ -282,24 +282,24 @@ class Labels {
 }
 
 class TimelineItem {
-  final String _type;
-  String get type => _type;
-  final String _title;
-  String get title => _title;
-  final int _number;
-  int get number => _number;
-  final Actor _actor;
-  Actor get actor => _actor;
-  final DateTime _createdAt;
-  DateTime get createdAt => _createdAt;
+  final String? _type;
+  String? get type => _type;
+  final String? _title;
+  String? get title => _title;
+  final int? _number;
+  int? get number => _number;
+  final Actor? _actor;
+  Actor? get actor => _actor;
+  final DateTime? _createdAt;
+  DateTime? get createdAt => _createdAt;
 
   TimelineItem(
       this._type, this._title, this._number, this._actor, this._createdAt);
 
   static TimelineItem fromGraphQL(dynamic node) {
-    String title = null;
-    int number = null;
-    Actor actor = null;
+    String? title = null;
+    int? number = null;
+    Actor? actor = null;
 
     if (node['__typename'] == 'MilestonedEvent' ||
         node['__typename'] == 'DemilestonedEvent') {
@@ -320,8 +320,8 @@ class TimelineItem {
 
   @override
   String toString() {
-    var result = '${_type} (' + _createdAt.toIso8601String() + ')';
-    result = '${result}' + (_actor != null ? ' by ${actor.login}' : '');
+    var result = '${_type} (' + _createdAt!.toIso8601String() + ')';
+    result = '${result}' + (_actor != null ? ' by ${actor!.login}' : '');
 
     if (type == 'CrossReferencedEvent') {
       result =
@@ -331,9 +331,9 @@ class TimelineItem {
     } else if (_type == 'DemilestonedEvent') {
       result = '${result} < ${title}';
     } else if (_type == 'AssignedEvent') {
-      result = '${result} > ${actor.login}';
+      result = '${result} > ${actor!.login}';
     } else if (_type == 'UnassignedEvent') {
-      result = '${result} < ${actor.login}';
+      result = '${result} < ${actor!.login}';
     }
 
     return result;
@@ -484,21 +484,21 @@ class Timeline {
 }
 
 class Milestone {
-  final String _title;
+  final String? _title;
   get title => _title;
-  final String _id;
+  final String? _id;
   get id => _id;
-  final int _number;
+  final int? _number;
   get number => _number;
-  final String _url;
+  final String? _url;
   get url => _url;
-  final bool _closed;
+  final bool? _closed;
   get closed => closed;
-  final DateTime _createdAt;
+  final DateTime? _createdAt;
   get createdAt => _createdAt;
-  final DateTime _closedAt;
+  final DateTime? _closedAt;
   get closedAt => _closedAt;
-  final DateTime _dueOn;
+  final DateTime? _dueOn;
   get dueOn => _dueOn;
 
   Milestone(this._title, this._id, this._number, this._url, this._closed,
@@ -573,21 +573,21 @@ class Repository {
 }
 
 class Actor {
-  final String _login;
+  final String? _login;
   get login => _login;
-  final String _url;
+  final String? _url;
   get url => _url;
 
   Actor(this._login, this._url);
-  static Actor fromGraphQL(dynamic node) {
+  static Actor? fromGraphQL(dynamic node) {
     return node != null && node['login'] != null
         ? Actor(node['login'], node['url'])
         : null;
   }
 
   @override
-  String toString() => this._login;
-  String toCsv() => this._login;
+  String toString() => this._login!;
+  String? toCsv() => this._login;
 
   @override
   bool operator ==(Object other) =>
@@ -608,24 +608,24 @@ class Actor {
 }
 
 class Organization {
-  final String _id;
+  final String? _id;
   get id => _id;
-  final String _avatarUrl;
+  final String? _avatarUrl;
   get avatarUrl => _avatarUrl;
-  final DateTime _createdAt;
+  final DateTime? _createdAt;
   get createdAt => _createdAt;
-  final String _description;
+  final String? _description;
   get description => _description;
-  final String _email;
+  final String? _email;
   get email => _email;
-  final String _login;
+  final String? _login;
   get login => _login;
-  final String _name;
+  final String? _name;
   get name => _name;
 
   get pendingMembersStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = Organization.request(id, pendingMembersAfter: after);
       final options = QueryOptions(document: gql(query));
@@ -633,16 +633,16 @@ class Organization {
       final page = await _client.query(options);
       try {
         PageInfo pageInfo = PageInfo.fromGraphQL(
-            page.data['organization']['pendingMembers']['pageInfo']);
+            page.data!['organization']['pendingMembers']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
         return;
       }
       // Parse the responses into a buffer
-      var membersBuffer = <Actor>[];
+      var membersBuffer = <Actor?>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['organization']['pendingMembers']
+      for (var jsonSub in page.data!['organization']['pendingMembers']
           ['edges']) {
         membersBuffer.add(Actor.fromGraphQL(jsonSub['node']));
       }
@@ -652,19 +652,19 @@ class Organization {
         do {
           yield membersBuffer[bufferIndex++];
         } while (bufferIndex < membersBuffer.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   get teamsStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
-      var query = Organization.request(_login, teamsAfter: after);
+      var query = Organization.request(_login!, teamsAfter: after);
       final options = QueryOptions(document: gql(query));
       final page = await _client.query(options);
       try {
         PageInfo pageInfo = PageInfo.fromGraphQL(
-            page.data['organization']['teams']['pageInfo']);
+            page.data!['organization']['teams']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
@@ -673,7 +673,7 @@ class Organization {
       // Parse the responses into a buffer
       var teamsBuffer = <Team>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['organization']['teams']['edges']) {
+      for (var jsonSub in page.data!['organization']['teams']['edges']) {
         teamsBuffer.add(Team.fromGraphQL(jsonSub['node']));
       }
 
@@ -682,7 +682,7 @@ class Organization {
         do {
           yield teamsBuffer[bufferIndex++];
         } while (bufferIndex < teamsBuffer.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   Organization(this._id, this._avatarUrl, this._createdAt, this._description,
@@ -700,9 +700,9 @@ class Organization {
   }
 
   static String request(String login,
-      {String pendingMembersAfter = null,
-      String repositoriesAfter = null,
-      String teamsAfter = null}) {
+      {String? pendingMembersAfter = null,
+      String? repositoriesAfter = null,
+      String? teamsAfter = null}) {
     return Organization._childQuery
         .replaceAll(r'${login}', login)
         .replaceAll(
@@ -778,22 +778,22 @@ class Organization {
 }
 
 class Team {
-  final String _id;
+  final String? _id;
   get id => _id;
-  final String _avatarUrl;
+  final String? _avatarUrl;
   get avatarUrl => _avatarUrl;
-  final String _name;
+  final String? _name;
   get name => _name;
-  final String _description;
+  final String? _description;
   get description => _description;
-  final DateTime _createdAt;
+  final DateTime? _createdAt;
   get createdAt => _createdAt;
-  DateTime _updatedAt;
+  DateTime? _updatedAt;
   get updatedAt => _updatedAt;
 
   get membersStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = Team.request(id, membersAfter: after);
       final options = QueryOptions(document: gql(query));
@@ -801,16 +801,16 @@ class Team {
       final page = await _client.query(options);
       try {
         PageInfo pageInfo =
-            PageInfo.fromGraphQL(page.data['node']['members']['pageInfo']);
+            PageInfo.fromGraphQL(page.data!['node']['members']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
         return;
       }
       // Parse the responses into a buffer
-      var membersBuffer = <Actor>[];
+      var membersBuffer = <Actor?>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['node']['members']['edges']) {
+      for (var jsonSub in page.data!['node']['members']['edges']) {
         membersBuffer.add(Actor.fromGraphQL(jsonSub['node']));
       }
 
@@ -819,12 +819,12 @@ class Team {
         do {
           yield membersBuffer[bufferIndex++];
         } while (bufferIndex < membersBuffer.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   get childTeamsStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = Team.request(id, childTeamsAfter: after);
       final options = QueryOptions(document: gql(query));
@@ -832,7 +832,7 @@ class Team {
       final page = await _client.query(options);
       try {
         PageInfo pageInfo =
-            PageInfo.fromGraphQL(page.data['node']['childTeams']['pageInfo']);
+            PageInfo.fromGraphQL(page.data!['node']['childTeams']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
@@ -841,7 +841,7 @@ class Team {
       // Parse the responses into a buffer
       var teamBuffer = <Team>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['node']['childTeams']['edges']) {
+      for (var jsonSub in page.data!['node']['childTeams']['edges']) {
         teamBuffer.add(Team.fromGraphQL(jsonSub));
       }
 
@@ -850,7 +850,7 @@ class Team {
         do {
           yield teamBuffer[bufferIndex++];
         } while (bufferIndex < teamBuffer.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   Team(this._id, this._avatarUrl, this._createdAt, this._description,
@@ -867,7 +867,7 @@ class Team {
   }
 
   static String request(String id,
-      {String childTeamsAfter = null, String membersAfter = null}) {
+      {String? childTeamsAfter = null, String? membersAfter = null}) {
     return Team._childQuery
         .replaceAll(r'${ownerId}', id)
         .replaceAll(r'${childTeamsAfter}',
@@ -931,41 +931,41 @@ class Team {
 }
 
 class Issue {
-  final String _title;
+  final String? _title;
   get title => _title;
-  final String _id;
+  final String? _id;
   get id => _id;
-  final int _number;
+  final int? _number;
   get number => _number;
-  final String _state;
+  final String? _state;
   get state => _state;
-  final Actor _author;
+  final Actor? _author;
   get author => _author;
-  final List<Actor> _assignees;
+  final List<Actor?>? _assignees;
   get assignees => _assignees;
-  final String _body;
+  final String? _body;
   get body => _body;
-  final Labels _labels;
+  final Labels? _labels;
   get labels => _labels;
-  final String _url;
+  final String? _url;
   get url => _url;
-  final DateTime _createdAt;
+  final DateTime? _createdAt;
   get createdAt => _createdAt;
-  final DateTime _closedAt;
+  final DateTime? _closedAt;
   get closedAt => _closedAt;
-  final DateTime _lastEditAt;
+  final DateTime? _lastEditAt;
   get lastEditAt => _lastEditAt;
-  final DateTime _updatedAt;
+  final DateTime? _updatedAt;
   get updatedAt => _updatedAt;
-  final Repository _repository;
+  final Repository? _repository;
   get repository => _repository;
-  final Milestone _milestone;
+  final Milestone? _milestone;
   get milestone => _milestone;
-  final Timeline _timeline;
+  final Timeline? _timeline;
   get timeline => _timeline;
   get reactionStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = _reactionQuery
           .replaceAll(r'${issue}', this.number.toString())
@@ -975,7 +975,7 @@ class Issue {
 
       try {
         PageInfo pageInfo = PageInfo.fromGraphQL(
-            page.data['repository']['issue']['reactions']['pageInfo']);
+            page.data!['repository']['issue']['reactions']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
@@ -984,7 +984,7 @@ class Issue {
       // Parse the responses into a buffer
       var bufferReactions = <Reaction>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['repository']['issue']['reactions']
+      for (var jsonSub in page.data!['repository']['issue']['reactions']
           ['nodes']) {
         bufferReactions.add(Reaction.fromGraphQL(jsonSub));
       }
@@ -994,22 +994,22 @@ class Issue {
         do {
           yield bufferReactions[bufferIndex++];
         } while (bufferIndex < bufferReactions.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   get commentStream async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = _commentQuery
-          .replaceAll(r'${ownerId}', _id)
+          .replaceAll(r'${ownerId}', _id!)
           .replaceAll(r'${after}', after);
       final options = QueryOptions(document: gql(query));
 
       final page = await _client.query(options);
       try {
         PageInfo pageInfo =
-            PageInfo.fromGraphQL(page.data['node']['comments']['pageInfo']);
+            PageInfo.fromGraphQL(page.data!['node']['comments']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
@@ -1018,7 +1018,7 @@ class Issue {
       // Parse the responses into a buffer
       var commentBuffer = <Comment>[];
       var bufferIndex = 0;
-      for (var jsonSub in page.data['node']['comments']['nodes']) {
+      for (var jsonSub in page.data!['node']['comments']['nodes']) {
         commentBuffer.add(Comment.fromGraphQL(jsonSub));
       }
 
@@ -1027,7 +1027,7 @@ class Issue {
         do {
           yield commentBuffer[bufferIndex++];
         } while (bufferIndex < commentBuffer.length);
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   Issue(
@@ -1050,10 +1050,10 @@ class Issue {
 
   // Passed a node containing an issue, return the issue
   static Issue fromGraphQL(dynamic node) {
-    List<Actor> assignees = null;
+    List<Actor?>? assignees = null;
     var edges = (node['assignees'] ?? {})['edges'];
     if (edges != null && edges.length != 0) {
-      assignees = <Actor>[];
+      assignees = <Actor?>[];
       for (var node in edges) {
         assignees.add(Actor.fromGraphQL(node['node']));
       }
@@ -1097,37 +1097,38 @@ class Issue {
       showMilestone = false,
       bool linebreakAfter = false,
       includeLabels: true}) {
-    var labelsSummary = includeLabels ? _labels.summary() : '';
+    var labelsSummary = includeLabels ? _labels!.summary() : '';
     var markdown = '[${this.number}](${this.url})';
     markdown = '${markdown} ${this.title} ${labelsSummary}';
     if (showMilestone) {
       markdown = '${markdown} ' +
-          (_milestone == null ? '[no milestone]' : '[${_milestone.title}]');
+          (_milestone == null ? '[no milestone]' : '[${_milestone!.title}]');
     }
-    if (boldInteresting && _labels.intersect(_interesting))
+    if (boldInteresting && _labels!.intersect(_interesting))
       markdown = '**' + markdown + '**';
     if (linebreakAfter) markdown = markdown + '\n';
     return markdown;
   }
 
   String verbose({bool boldInteresting = true, bool linebreakAfter = false}) {
-    var labelsSummary = _labels.summary();
+    var labelsSummary = _labels!.summary();
     var markdown = '[${this.number}](${this.url})';
-    if (_assignees == null || _assignees.length == 0)
+    if (_assignees == null || _assignees!.length == 0)
       markdown = '${markdown} > UNASSIGNED';
     else {
       markdown = '${markdown} > (';
-      _assignees
-          .forEach((assignee) => markdown = '${markdown}${assignee.login}, ');
+      _assignees!
+          .forEach((assignee) => markdown = '${markdown}${assignee!.login}, ');
       markdown = markdown.substring(0, markdown.length - 2);
       markdown = '${markdown})';
     }
     if (_milestone == null)
       markdown = '${markdown} with no milestone';
     else
-      markdown = '${markdown} due on ${_milestone.dueOn} (${_milestone.title})';
+      markdown =
+          '${markdown} due on ${_milestone!.dueOn} (${_milestone!.title})';
     markdown = '${markdown} ${this.title} ${labelsSummary}';
-    if (boldInteresting && _labels.intersect(_interesting))
+    if (boldInteresting && _labels!.intersect(_interesting))
       markdown = '**' + markdown + '**';
     if (linebreakAfter) markdown = markdown + '\n';
     return markdown;
@@ -1140,38 +1141,38 @@ class Issue {
   // and Google Sheets only takes mixed CSV/TSV records with TSV being the containing
   // record format.
   String toTsv() {
-    String milestoneHistory = '';
+    String? milestoneHistory = '';
     if (timeline != null)
       timeline.milestoneTimeline.forEach((milestone) => milestoneHistory =
           milestone == null
               ? milestoneHistory
               : '${milestoneHistory},${milestone.title}');
-    if (milestoneHistory.length > 0)
-      milestoneHistory = milestoneHistory.substring(1);
-    if (milestoneHistory.length == 0)
-      milestoneHistory = _milestone != null ? _milestone.title : '';
+    if (milestoneHistory!.length > 0)
+      milestoneHistory = milestoneHistory!.substring(1);
+    if (milestoneHistory!.length == 0)
+      milestoneHistory = _milestone != null ? _milestone!.title : '';
 
     var originalMilestone;
     if (_timeline == null) {
       originalMilestone = '';
     } else {
-      if (_timeline.originalMilestone == null) {
+      if (_timeline!.originalMilestone == null) {
         originalMilestone = '';
       } else {
-        _timeline.originalMilestone.title;
+        _timeline!.originalMilestone.title;
       }
     }
-    var currentMilestone = _milestone == null ? '' : _milestone.title;
-    var dueOn = _milestone == null ? '' : _milestone.dueOn.toString();
+    var currentMilestone = _milestone == null ? '' : _milestone!.title;
+    var dueOn = _milestone == null ? '' : _milestone!.dueOn.toString();
 
     String tsv = '';
     tsv = '${tsv}=HYPERLINK("${_url}","${_number}")';
     tsv = '${tsv}\t${_title}';
-    tsv = '${tsv}\t${_labels.priority()}';
+    tsv = '${tsv}\t${_labels!.priority()}';
     tsv = '${tsv}\t${_state}';
-    tsv = '${tsv}\t' + (_author == null ? '' : _author.toCsv());
+    tsv = '${tsv}\t' + (_author == null ? '' : _author!.toCsv()!);
     tsv = '${tsv}\t${createdAt}';
-    if (_assignees != null && _assignees.length > 0) {
+    if (_assignees != null && _assignees!.length > 0) {
       tsv = '${tsv}\t';
       assignees.forEach((assignee) => tsv = '${tsv}${assignee.login},');
       tsv = tsv.substring(0, tsv.length - 1);
@@ -1305,43 +1306,43 @@ class Issue {
 }
 
 class PullRequest {
-  final String _title;
+  final String? _title;
   get title => _title;
-  final String _id;
+  final String? _id;
   get id => _id;
-  final int _number;
+  final int? _number;
   get number => _number;
-  final String _state;
+  final String? _state;
   get state => _state;
-  final Actor _author;
+  final Actor? _author;
   get author => _author;
-  final List<Actor> _reviewers;
+  final List<Actor?>? _reviewers;
   get reviewers => _reviewers;
-  final List<Actor> _assignees;
+  final List<Actor?>? _assignees;
   get assignees => _assignees;
-  final String _body;
+  final String? _body;
   get body => _body;
-  final Milestone _milestone;
+  final Milestone? _milestone;
   get milestone => _milestone;
-  final Labels _labels;
+  final Labels? _labels;
   get labels => _labels;
-  final String _url;
+  final String? _url;
   get url => _url;
-  final bool _merged;
+  final bool? _merged;
   get merged => _merged;
-  final DateTime _createdAt;
+  final DateTime? _createdAt;
   get createdAt => _createdAt;
-  final DateTime _mergedAt;
+  final DateTime? _mergedAt;
   get mergedAt => _mergedAt;
-  final DateTime _lastEditAt;
+  final DateTime? _lastEditAt;
   get lastEditAt => _lastEditAt;
-  final DateTime _updatedAt;
+  final DateTime? _updatedAt;
   get updatedAt => _updatedAt;
-  final DateTime _closedAt;
+  final DateTime? _closedAt;
   get closedAt => _closedAt;
-  final Repository _repository;
+  final Repository? _repository;
   get repository => _repository;
-  final Timeline _timeline;
+  final Timeline? _timeline;
   get timeline => _timeline;
 
   PullRequest(
@@ -1367,18 +1368,18 @@ class PullRequest {
 
   // Passed a node containing an issue, return the issue
   static PullRequest fromGraphQL(dynamic node) {
-    List<Actor> assignees = null;
-    List<Actor> reviewers = null;
+    List<Actor?>? assignees = null;
+    List<Actor?>? reviewers = null;
     if (node['assignees']['edges'] != null &&
         node['assignees']['edges'].length != 0) {
-      assignees = <Actor>[];
+      assignees = <Actor?>[];
       for (var node in node['assignees']['edges']) {
         assignees.add(Actor.fromGraphQL(node['node']));
       }
     }
     if (node['reviews']['edges'] != null &&
         node['reviews']['edges'].length != 0) {
-      reviewers = <Actor>[];
+      reviewers = <Actor?>[];
       for (var node in node['reviews']['edges']) {
         if (node['node']['author'] != null)
           reviewers.add(Actor.fromGraphQL(node['node']['author']));
@@ -1418,7 +1419,7 @@ class PullRequest {
       {bool linebreakAfter = false,
       bool boldInteresting = false,
       includeLabels: true}) {
-    var labelsSummary = includeLabels ? _labels.summary() : '';
+    var labelsSummary = includeLabels ? _labels!.summary() : '';
     var markdown =
         '[${this.number}](${this.url}) ${this.title} ${labelsSummary}';
     if (linebreakAfter) markdown = markdown + '\n';
@@ -1426,21 +1427,22 @@ class PullRequest {
   }
 
   String verbose({bool boldInteresting = true, bool linebreakAfter = false}) {
-    var labelsSummary = _labels.summary();
+    var labelsSummary = _labels!.summary();
     var markdown = '[${this.number}](${this.url})';
-    if (_assignees == null || _assignees.length == 0)
+    if (_assignees == null || _assignees!.length == 0)
       markdown = '${markdown} > UNASSIGNED';
     else {
       markdown = '${markdown} > (';
-      _assignees
-          .forEach((assignee) => markdown = '${markdown}${assignee.login}, ');
+      _assignees!
+          .forEach((assignee) => markdown = '${markdown}${assignee!.login}, ');
       markdown = markdown.substring(0, markdown.length - 2);
       markdown = '${markdown})';
     }
     if (_milestone == null)
       markdown = '${markdown} with no milestone';
     else
-      markdown = '${markdown} due on ${_milestone.dueOn} (${_milestone.title})';
+      markdown =
+          '${markdown} due on ${_milestone!.dueOn} (${_milestone!.title})';
     markdown = '${markdown} ${this.title} ${labelsSummary}';
     if (linebreakAfter) markdown = markdown + '\n';
     return markdown;
@@ -1453,44 +1455,44 @@ class PullRequest {
   // and Google Sheets only takes mixed CSV/TSV records with TSV being the containing
   // record format.
   String toTsv() {
-    var milestoneHistory = '';
+    String? milestoneHistory = '';
     if (timeline != null)
       timeline.milestoneTimeline().forEach((milestone) =>
           milestoneHistory = '${milestoneHistory},${milestone.title}');
-    if (milestoneHistory.length > 0)
-      milestoneHistory = milestoneHistory.substring(1);
-    if (milestoneHistory.length == 0)
+    if (milestoneHistory!.length > 0)
+      milestoneHistory = milestoneHistory!.substring(1);
+    if (milestoneHistory!.length == 0)
       milestoneHistory = _milestone != null ? milestone.title : '';
     var originalMilestone;
     if (_timeline == null) {
       originalMilestone = '';
     } else {
-      if (_timeline.originalMilestone == null) {
+      if (_timeline!.originalMilestone == null) {
         originalMilestone = '';
       } else {
-        _timeline.originalMilestone.title;
+        _timeline!.originalMilestone.title;
       }
     }
-    var currentMilestone = _milestone == null ? '' : _milestone.title;
-    var dueOn = _milestone == null ? '' : _milestone.dueOn.toString();
+    var currentMilestone = _milestone == null ? '' : _milestone!.title;
+    var dueOn = _milestone == null ? '' : _milestone!.dueOn.toString();
 
     String tsv = '';
     tsv = '${tsv}=HYPERLINK("${_url}","${_number}")';
     tsv = '${tsv}\t${_title}';
-    tsv = '${tsv}\t${_labels.priority()}';
-    tsv = '${tsv}\t${_author.toCsv()}';
+    tsv = '${tsv}\t${_labels!.priority()}';
+    tsv = '${tsv}\t${_author!.toCsv()}';
     tsv = '${tsv}\t${createdAt}';
-    tsv = '${tsv}\t' + (_merged ? 'Y' : 'N');
-    if (_assignees != null && _assignees.length > 0) {
+    tsv = '${tsv}\t' + (_merged! ? 'Y' : 'N');
+    if (_assignees != null && _assignees!.length > 0) {
       tsv = '${tsv}\t';
-      _assignees.forEach((assignee) => tsv = '${tsv}${assignee.login},');
+      _assignees!.forEach((assignee) => tsv = '${tsv}${assignee!.login},');
       tsv = tsv.substring(0, tsv.length - 1);
     } else {
       tsv = '${tsv}\t';
     }
-    if (_reviewers != null && _reviewers.length > 0) {
+    if (_reviewers != null && _reviewers!.length > 0) {
       tsv = '${tsv}\t';
-      _reviewers.forEach((reviewer) => tsv = '${tsv}${reviewer.login},');
+      _reviewers!.forEach((reviewer) => tsv = '${tsv}${reviewer!.login},');
       tsv = tsv.substring(0, tsv.length - 1);
     } else {
       tsv = '${tsv}\t';
@@ -1567,7 +1569,7 @@ enum ClusterReportSort { byKey, byCount }
 class Cluster {
   final ClusterType _type;
   get type => _type;
-  final SplayTreeMap<String, dynamic> _clusters;
+  final SplayTreeMap<String?, dynamic> _clusters;
   get clusters => _clusters;
   get keys => _clusters.keys;
   dynamic operator [](String key) => _clusters[key];
@@ -1581,12 +1583,12 @@ class Cluster {
   static final _noMilestoneKey = '__no milestone__';
 
   static Cluster byLabel(List<dynamic> issuesOrPullRequests) {
-    var result = SplayTreeMap<String, dynamic>();
+    var result = SplayTreeMap<String?, dynamic>();
     result[_unlabeledKey] = [];
 
     for (var item in issuesOrPullRequests) {
       if (!(item is Issue) && !(item is PullRequest)) {
-        throw ('invalid type!');
+        throw ArgumentError('invalid type!');
       }
       if (item.labels != null) {
         for (var label in item.labels.labels) {
@@ -1605,11 +1607,11 @@ class Cluster {
   }
 
   static Cluster byAuthor(List<dynamic> issuesOrPullRequests) {
-    var result = SplayTreeMap<String, dynamic>();
+    var result = SplayTreeMap<String?, dynamic>();
 
     for (var item in issuesOrPullRequests) {
       if (!(item is Issue) && !(item is PullRequest)) {
-        throw ('invalid type!');
+        throw ArgumentError('invalid type!');
       }
       var name = item.author != null ? item.author.login : '@@@ NO AUTHOR @@@';
       if (!result.containsKey(name)) {
@@ -1622,12 +1624,12 @@ class Cluster {
   }
 
   static Cluster byAssignees(List<dynamic> issuesOrPullRequests) {
-    var result = SplayTreeMap<String, dynamic>();
+    var result = SplayTreeMap<String?, dynamic>();
     result[_unassignedKey] = [];
 
     for (var item in issuesOrPullRequests) {
       if (!(item is Issue) && !(item is PullRequest)) {
-        throw ('invalid type!');
+        throw ArgumentError('invalid type!');
       }
       if (item.assignees == null || item.assignees.length == 0) {
         result[_unassignedKey].add(item);
@@ -1645,14 +1647,14 @@ class Cluster {
   }
 
   static Cluster byReviewers(List<dynamic> issuesOrPullRequests) {
-    var result = SplayTreeMap<String, dynamic>();
+    var result = SplayTreeMap<String?, dynamic>();
     result[_unassignedKey] = [];
 
     for (var item in issuesOrPullRequests) {
       if (!(item is PullRequest)) {
-        throw ('invalid type!');
+        throw ArgumentError('invalid type!');
       }
-      var pr = item as PullRequest;
+      var pr = item;
       if (pr.reviewers == null || pr.reviewers.length == 0) {
         result[_unassignedKey].add(item);
       } else
@@ -1669,12 +1671,12 @@ class Cluster {
   }
 
   static Cluster byMilestone(List<dynamic> issuesOrPullRequests) {
-    var result = SplayTreeMap<String, dynamic>();
+    var result = SplayTreeMap<String?, dynamic>();
     result[_noMilestoneKey] = [];
 
     for (var item in issuesOrPullRequests) {
       if (!(item is Issue) && !(item is PullRequest)) {
-        throw ('invalid type!');
+        throw ArgumentError('invalid type!');
       }
       if (item.milestone == null) {
         result[_noMilestoneKey].add(item);
@@ -1716,7 +1718,7 @@ class Cluster {
   String toString() => summary();
 
   String toMarkdown(
-      {ClusterReportSort sortType,
+      {ClusterReportSort? sortType,
       bool skipEmpty = true,
       showStatistics = false}) {
     var result = '';
@@ -1775,7 +1777,7 @@ class Cluster {
 
   double mean() {
     double sum = 0.0;
-    double l = 0.0;
+    double? l = 0.0;
     switch (type) {
       case ClusterType.byAuthor:
         l = (clusters.keys.contains(_unassignedKey)
@@ -1804,13 +1806,13 @@ class Cluster {
         break;
     }
     clusters.keys.forEach((key) => sum += clusters[key].length);
-    return sum / l;
+    return sum / l!;
   }
 
   double stdev() {
     double m = mean();
     double sum = 0.0;
-    double l = 0.0;
+    double? l = 0.0;
     switch (type) {
       case ClusterType.byAuthor:
         l = (clusters.keys.contains(_unassignedKey)
@@ -1841,7 +1843,7 @@ class Cluster {
     clusters.keys.forEach((key) =>
         sum += ((clusters[key].length - m) * (clusters[key].length - m)));
 
-    double deviation = sum / l;
+    double deviation = sum / l!;
 
     return sqrt(deviation);
   }
