@@ -7,16 +7,16 @@ enum GitHubDateQueryType { created, updated, closed, merged, none }
 
 /// Used to perform queries against GitHub.
 class GitHub {
-  HttpLink _httpLink;
-  AuthLink _auth;
-  Link _link;
-  GraphQLClient _client;
+  late HttpLink _httpLink;
+  late AuthLink _auth;
+  late Link _link;
+  late GraphQLClient _client;
 
   var _maxSearchResponse = 1000;
   var _printQuery = false;
 
   /// Initialize the interface
-  GitHub(String token) {
+  GitHub(String? token) {
     _httpLink = HttpLink('https://api.github.com/graphql');
     _auth = AuthLink(getToken: () async => 'Bearer ${token}');
     _link = _auth.concat(_httpLink);
@@ -31,9 +31,9 @@ class GitHub {
     final page = await _client.query(options);
     if (page.hasException) {
       print(query);
-      throw page.exception;
+      throw page.exception!;
     }
-    return Team.fromGraphQL(page.data['node']);
+    return Team.fromGraphQL(page.data!['node']);
   }
 
   /// Fetch an organiation by its login.
@@ -44,9 +44,9 @@ class GitHub {
     final page = await _client.query(options);
     if (page.hasException) {
       print(query);
-      throw page.exception;
+      throw page.exception!;
     }
-    return Organization.fromGraphQL(page.data['organization']);
+    return Organization.fromGraphQL(page.data!['organization']);
   }
 
   ///
@@ -56,7 +56,7 @@ class GitHub {
   /// Returns a string of PullRequest and Issue items.
   Stream<dynamic> searchIssuePRs(String queryString) async* {
     var after = 'null';
-    bool hasNextPage;
+    bool? hasNextPage;
     do {
       var query = _searchIssuesOrPRs
           .replaceAll(r'${query}', queryString.replaceAll('"', '\\"'))
@@ -71,12 +71,12 @@ class GitHub {
         print(query);
         print(page.exception);
         print(page.data);
-        throw page.exception;
+        throw page.exception!;
       }
       // Paginate information
       try {
         PageInfo pageInfo =
-            PageInfo.fromGraphQL(page.data['search']['pageInfo']);
+            PageInfo.fromGraphQL(page.data!['search']['pageInfo']);
         hasNextPage = pageInfo.hasNextPage;
         after = '"${pageInfo.endCursor}"';
       } on Error {
@@ -85,7 +85,7 @@ class GitHub {
 
       var buffer = [];
       var bufferIndex = 0;
-      var edges = page.data['search']['nodes'];
+      var edges = page.data!['search']['nodes'];
       edges.forEach((edge) {
         dynamic item = edge['__typename'] == 'Issue'
             ? Issue.fromGraphQL(edge)
@@ -97,7 +97,7 @@ class GitHub {
           yield buffer[bufferIndex++];
         } while (bufferIndex < buffer.length);
       }
-    } while (hasNextPage);
+    } while (hasNextPage!);
   }
 
   /// Search for issues and PRs matching criteria across a date range.
@@ -108,13 +108,13 @@ class GitHub {
   /// but only a count.
   /// DEPRECATED
   Future<List<dynamic>> deprecated_search(
-      {String owner,
-      String name,
+      {String? owner,
+      String? name,
       GitHubIssueType type = GitHubIssueType.issue,
       GitHubIssueState state = GitHubIssueState.open,
-      List<String> labels = null,
+      List<String>? labels = null,
       GitHubDateQueryType dateQuery = GitHubDateQueryType.none,
-      DateRange dateRange = null}) async {
+      DateRange? dateRange = null}) async {
     var typeString = type == GitHubIssueType.issue ? 'issue' : 'pr';
     var stateString = '';
     switch (state) {
@@ -150,18 +150,18 @@ class GitHub {
 
     var fetchAnotherDay = false;
     var splitFetches = false;
-    var totalIssueCount = null;
+    dynamic totalIssueCount = null;
     var result = [];
-    var resultsFetched = Set<int>();
+    var resultsFetched = Set<int?>();
     // For each label, do the query.
     for (var labelFilter in labelFilters) {
       do {
-        var fetchAnotherPage = false;
+        bool? fetchAnotherPage = false;
         var after = 'null';
         do {
           var query = _deprecated_searchIssuesOrPRs
-              .replaceAll(r'${repositoryOwner}', owner)
-              .replaceAll(r'${repositoryName}', name)
+              .replaceAll(r'${repositoryOwner}', owner!)
+              .replaceAll(r'${repositoryName}', name!)
               .replaceAll(r'${after}', after)
               .replaceAll(r'${label}', labelFilter)
               .replaceAll(r'${state}', stateString)
@@ -176,9 +176,9 @@ class GitHub {
           final page = await _client.query(options);
           if (page.hasException) {
             print(query);
-            throw page.exception;
+            throw page.exception!;
           }
-          var edges = page.data['search']['nodes'];
+          var edges = page.data!['search']['nodes'];
           edges.forEach((edge) {
             dynamic item = type == GitHubIssueType.issue
                 ? Issue.fromGraphQL(edge)
@@ -192,16 +192,16 @@ class GitHub {
 
           // Pseudo-paginate by date if we're not already, and there's too many responses.
           if (!splitFetches &&
-              page.data['search']['issueCount'] >= _maxSearchResponse) {
+              page.data!['search']['issueCount'] >= _maxSearchResponse) {
             splitFetches = true;
           }
 
           // GitHub pagination
           if (totalIssueCount == null)
-            totalIssueCount = page.data['search']['issueCount'];
-          var pageInfo = PageInfo.fromGraphQL(page.data['search']['pageInfo']);
+            totalIssueCount = page.data!['search']['issueCount'];
+          var pageInfo = PageInfo.fromGraphQL(page.data!['search']['pageInfo']);
           fetchAnotherPage = pageInfo.hasNextPage;
-          if (fetchAnotherPage) after = '"${pageInfo.endCursor}"';
+          if (fetchAnotherPage!) after = '"${pageInfo.endCursor}"';
         } while (fetchAnotherPage);
 
         // pseudo-pagination -- if this response returns its maximum
@@ -210,20 +210,20 @@ class GitHub {
         // If we need to split fetches across days, do so.
         if (splitFetches) {
           fetchAnotherDay = true;
-          switch (dateRange.type) {
+          switch (dateRange!.type) {
             case DateRangeType.at:
-              throw ('unsupported DateRangeType.at with maximum number of elements');
-              break;
+              throw ArgumentError(
+                  'unsupported DateRangeType.at with maximum number of elements');
             case DateRangeType.range:
               final dayDelta = 4;
-              var newEnd = startSearchFrom.end;
+              var newEnd = startSearchFrom!.end;
               startSearchFrom = DateRange(DateRangeType.range,
                   start: newEnd.subtract(Duration(days: 2 * dayDelta)),
                   end: newEnd.subtract(Duration(days: dayDelta)));
               var newDateRange = DateRange(DateRangeType.range,
                   start: newEnd
                           .subtract(Duration(days: dayDelta))
-                          .isBefore(endSearchAt.start)
+                          .isBefore(endSearchAt!.start)
                       ? endSearchAt.start
                       : newEnd.subtract(Duration(days: dayDelta)),
                   end: newEnd);
@@ -256,13 +256,13 @@ class GitHub {
   /// be used when the number of returned values is expected to be
   /// low.
   Future<List<dynamic>> fetch(
-      {String owner,
-      String name,
-      GitHubIssueType type = GitHubIssueType.issue,
+      {String? owner,
+      String? name,
+      GitHubIssueType? type = GitHubIssueType.issue,
       GitHubIssueState state = GitHubIssueState.open,
-      List<String> labels = null,
+      List<String>? labels = null,
       GitHubDateQueryType dateQuery = GitHubDateQueryType.none,
-      DateRange dateRange = null}) async {
+      DateRange? dateRange = null}) async {
     var typeString = type == GitHubIssueType.issue ? 'issues' : 'pullRequests';
     var stateString = '';
     switch (state) {
@@ -278,7 +278,8 @@ class GitHub {
     }
 
     if (dateQuery != GitHubDateQueryType.none && dateRange == null) {
-      throw ('With a dateQuery you must provide a non-null dateRange!');
+      throw ArgumentError(
+          'With a dateQuery you must provide a non-null dateRange!');
     }
 
     var result = [];
@@ -286,8 +287,8 @@ class GitHub {
     var after = 'null';
     do {
       var query = _queryIssuesOrPRs
-          .replaceAll(r'${repositoryOwner}', owner)
-          .replaceAll(r'${repositoryName}', name)
+          .replaceAll(r'${repositoryOwner}', owner!)
+          .replaceAll(r'${repositoryName}', name!)
           .replaceAll(r'${type}', typeString)
           .replaceAll(r'${after}', after)
           .replaceAll(r'${state}', stateString)
@@ -305,10 +306,10 @@ class GitHub {
       if (page.hasException) {
         print(query);
         print(page.source);
-        throw page.exception;
+        throw page.exception!;
       }
 
-      var edges = page.data['repository'][typeString]['nodes'];
+      var edges = page.data!['repository'][typeString]['nodes'];
       edges.forEach((edge) {
         dynamic item = type == GitHubIssueType.issue
             ? Issue.fromGraphQL(edge)
@@ -319,21 +320,21 @@ class GitHub {
           switch (dateQuery) {
             case GitHubDateQueryType.created:
               add = (item.createdAt != null &&
-                      item.createdAt.isAfter(dateRange.start) &&
+                      item.createdAt.isAfter(dateRange!.start) &&
                       item.createdAt.isBefore(dateRange.end))
                   ? true
                   : false;
               break;
             case GitHubDateQueryType.updated:
               add = (item.updatedAt != null &&
-                      item.updatedAt.isAfter(dateRange.start) &&
+                      item.updatedAt.isAfter(dateRange!.start) &&
                       item.updatedAt.isBefore(dateRange.end))
                   ? true
                   : false;
               break;
             case GitHubDateQueryType.closed:
               add = (item.closedAt != null &&
-                      item.closedAt.isAfter(dateRange.start) &&
+                      item.closedAt.isAfter(dateRange!.start) &&
                       item.closedAt.isBefore(dateRange.end))
                   ? true
                   : false;
@@ -344,7 +345,7 @@ class GitHub {
               } else {
                 add = (item.merged &&
                         item.mergedAt != null &&
-                        item.closedAt.isAfter(dateRange.start) &&
+                        item.closedAt.isAfter(dateRange!.start) &&
                         item.closedAt.isBefore(dateRange.end))
                     ? true
                     : false;
@@ -360,8 +361,8 @@ class GitHub {
           result.add(item);
         }
       });
-      PageInfo pageInfo =
-          PageInfo.fromGraphQL(page.data['repository'][typeString]['pageInfo']);
+      PageInfo pageInfo = PageInfo.fromGraphQL(
+          page.data!['repository'][typeString]['pageInfo']);
 
       done = done || !pageInfo.hasNextPage;
       if (!done) after = '"${pageInfo.endCursor}"';
@@ -389,7 +390,8 @@ class GitHub {
   }
 
   /// Fetch a single issue.
-  Future<Issue /*!*/ > issue({String owner, String name, int number}) async {
+  Future<Issue> issue(
+      {required String owner, required String name, int? number}) async {
     var query = _query_issue
         .replaceAll(r'${repositoryOwner}', owner)
         .replaceAll(r'${repositoryName}', name)
@@ -399,15 +401,15 @@ class GitHub {
     final options = QueryOptions(document: gql(query));
     final page = await _client.query(options);
     if (page.hasException) {
-      throw page.exception;
+      throw page.exception!;
     }
 
-    return Issue.fromGraphQL(page.data['repository']['issue']);
+    return Issue.fromGraphQL(page.data!['repository']['issue']);
   }
 
   /// Fetch a single PR.
   Future<PullRequest> pullRequest(
-      {String owner, String name, int number}) async {
+      {required String owner, required String name, int? number}) async {
     var query = _query_pullRequest
         .replaceAll(r'${repositoryOwner}', owner)
         .replaceAll(r'${repositoryName}', name)
@@ -417,10 +419,10 @@ class GitHub {
     final options = QueryOptions(document: gql(query));
     final page = await _client.query(options);
     if (page.hasException) {
-      throw page.exception;
+      throw page.exception!;
     }
 
-    return PullRequest.fromGraphQL(page.data['repository']['pullRequest']);
+    return PullRequest.fromGraphQL(page.data!['repository']['pullRequest']);
   }
 
   final _queryIssuesOrPRs = r'''
@@ -487,9 +489,9 @@ enum DateRangeWhen { onDate, onOrBefore, onOrAfter }
 class DateRange {
   DateRangeType _type;
   get type => _type;
-  DateRangeWhen _when;
+  DateRangeWhen? _when;
   get when => _when;
-  DateTime _at, _start, _end;
+  DateTime? _at, _start, _end;
   get at => _at;
   get start => _start;
   get end => _end;
@@ -507,16 +509,18 @@ class DateRange {
         case DateRangeWhen.onOrAfter:
           comparison = '>=';
           break;
+        case null:
+          comparison = 'null';
       }
-      return comparison + _at.toIso8601String();
+      return comparison + _at!.toIso8601String();
     } else {
-      return _start.toIso8601String().replaceAll('.000', '') +
+      return _start!.toIso8601String().replaceAll('.000', '') +
           '..' +
-          _end.toIso8601String().replaceAll('.000', '');
+          _end!.toIso8601String().replaceAll('.000', '');
     }
   }
 
-  static String queryToString(GitHubDateQueryType dateQuery, DateRange range) {
+  static String queryToString(GitHubDateQueryType dateQuery, DateRange? range) {
     var dateString = '';
     switch (dateQuery) {
       case GitHubDateQueryType.created:
@@ -538,12 +542,11 @@ class DateRange {
   }
 
   factory DateRange(type,
-      {DateTime at,
+      {DateTime? at,
       DateRangeWhen when = DateRangeWhen.onDate,
-      DateTime start,
-      DateTime end}) {
+      DateTime? start,
+      DateTime? end}) {
     if (type == DateRangeType.at &&
-        when != null &&
         at != null &&
         start == null &&
         end == null) {
